+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [1.7.3] - 2023-01-04
 ### Changed
 * CI: suppress `npm audit` warnings for dev dependencies, due to CVE in
@@ -145,6 +154,7 @@ client library took over this package name. Thanks to
 ownership.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-node/compare/v1.7.3...HEAD
 [1.7.3]: https://github.com/DeepLcom/deepl-node/compare/v1.7.2...v1.7.3
 [1.7.2]: https://github.com/DeepLcom/deepl-node/compare/v1.7.1...v1.7.2
 [1.7.1]: https://github.com/DeepLcom/deepl-node/compare/v1.7.0...v1.7.1