+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+* Fixed proxy example code in README
+
 ## [1.8.0] - 2023-01-26
 ### Added
 * New languages available: Korean (`'ko'`) and Norwegian (bokmål) (`'nb'`). Add
@@ -164,6 +168,7 @@ client library took over this package name. Thanks to
 ownership.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-node/compare/v1.8.0...HEAD
 [1.8.0]: https://github.com/DeepLcom/deepl-node/compare/v1.7.5...v1.8.0
 [1.7.5]: https://github.com/DeepLcom/deepl-node/compare/v1.7.4...v1.7.5
 [1.7.4]: https://github.com/DeepLcom/deepl-node/compare/v1.7.3...v1.7.4