+++ CHANGELOG.md
@@ -5,12 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [1.9.0] - 2023-03-22
 ### Added
 * Added platform and node version information to the user-agent string that is sent with API calls, along with an opt-out.
 * Added method for applications that use this library to identify themselves in API requests they make.
 ### Fixed
 * Fixed proxy example code in README
+
 
 ## [1.8.0] - 2023-01-26
 ### Added
@@ -171,7 +172,7 @@ client library took over this package name. Thanks to
 ownership.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-node/compare/v1.8.0...HEAD
+[1.9.0]: https://github.com/DeepLcom/deepl-node/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/DeepLcom/deepl-node/compare/v1.7.5...v1.8.0
 [1.7.5]: https://github.com/DeepLcom/deepl-node/compare/v1.7.4...v1.7.5
 [1.7.4]: https://github.com/DeepLcom/deepl-node/compare/v1.7.3...v1.7.4